# swagger.yaml
openapi: 3.0.0
info:
  title: tour query api master 2
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:5000/api/v1/tours
  variables: {}
- url: http://localhost:5000/api/v1/users
  variables: {}
paths:
  /tours:
    post:
      tags:
      - tours
      summary: Create New Tour
      operationId: CreateNewTour
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateNewTourRequest'
              - example:
                  ratingsAverage: 3.9
                  ratingsQuantity: 69
                  images:
                  - tour-1-1.jpg
                  - tour-1-2.jpg
                  - tour-1-3.jpg
                  name: 'patel pav bhaji '
                  duration: 7
                  maxGroupSize: 25
                  difficulty: medium
                  price: 7356
                  summary: '  Fumio Kishida'
                  description: Japan is an island country in East Asia. It is situated in the northwest Pacific Ocean, and is bordered on the west by the Sea of Japan, while extending from the Sea of Okhotsk in the north toward the East China Sea, Philippine Sea, and Taiwan in the south.
                  imageCover: tour-1-cover.jpg
            example:
              ratingsAverage: 3.9
              ratingsQuantity: 69
              images:
              - tour-1-1.jpg
              - tour-1-2.jpg
              - tour-1-3.jpg
              name: 'patel pav bhaji '
              duration: 7
              maxGroupSize: 25
              difficulty: medium
              price: 7356
              summary: '  Fumio Kishida'
              description: Japan is an island country in East Asia. It is situated in the northwest Pacific Ocean, and is bordered on the west by the Sea of Japan, while extending from the Sea of Okhotsk in the north toward the East China Sea, Philippine Sea, and Taiwan in the south.
              imageCover: tour-1-cover.jpg
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: http://localhost:5000/api/v1
        variables: {}
    get:
      tags:
      - tours
      summary: Get tours data
      operationId: Gettoursdata
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: http://localhost:5000/api/v1
        variables: {}
  /tours/5c88fa8cf4afda39709c2955:
    get:
      tags:
      - tours
      summary: get single tour
      operationId: getsingletour
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: http://localhost:5000/api/v1
        variables: {}
    patch:
      tags:
      - tours
      summary: update tour
      operationId: updatetour
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/updatetourrequest'
              - example:
                  guides:
                  - 5c8a1dfa2f8fb814b56fa181
                  - 634fcb80b7a994199ce0a985
            example:
              guides:
              - 5c8a1dfa2f8fb814b56fa181
              - 634fcb80b7a994199ce0a985
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: http://localhost:5000/api/v1
        variables: {}
  /tours/6348ef27e3cb7c41e8d8fc7a:
    delete:
      tags:
      - tours
      summary: delete tour
      operationId: deletetour
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: http://localhost:5000/api/v1
        variables: {}
...
  /users:
    get:
      tags:
      - authentication
      summary: get all user
      operationId: getalluser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: http://localhost:5000/api/v1
        variables: {}
...
components:
  schemas:
    CreateNewTourRequest:
      title: CreateNewTourRequest
      required:
      - ratingsAverage
      - ratingsQuantity
      - images
      - name
      - duration
      - maxGroupSize
      - difficulty
      - price
      - summary
      - description
      - imageCover
      type: object
      properties:
        ratingsAverage:
          type: number
        ratingsQuantity:
          type: integer
          format: int32
        images:
          type: array
          items:
            type: string
          description: ''
        name:
          type: string
        duration:
          type: integer
          format: int32
        maxGroupSize:
          type: integer
          format: int32
        difficulty:
          type: string
        price:
          type: integer
          format: int32
        summary:
          type: string
        description:
          type: string
        imageCover:
          type: string
      example:
        ratingsAverage: 3.9
        ratingsQuantity: 69
        images:
        - tour-1-1.jpg
        - tour-1-2.jpg
        - tour-1-3.jpg
        name: 'patel pav bhaji '
        duration: 7
        maxGroupSize: 25
        difficulty: medium
        price: 7356
        summary: '  Fumio Kishida'
        description: Japan is an island country in East Asia. It is situated in the northwest Pacific Ocean, and is bordered on the west by the Sea of Japan, while extending from the Sea of Okhotsk in the north toward the East China Sea, Philippine Sea, and Taiwan in the south.
        imageCover: tour-1-cover.jpg
    updatetourrequest:
      title: updatetourrequest
      required:
      - guides
      type: object
      properties:
        guides:
          type: array
          items:
            type: string
          description: ''
      example:
        guides:
        - 5c8a1dfa2f8fb814b56fa181
        - 634fcb80b7a994199ce0a985
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...